+++ releases.md
@@ -5,6 +5,7 @@
   - Enable custom `depth:` and `filter:` options to `Sampler#track`.
   - Change default filter to no-op.
   - Add option to run GC with custom options before each sample to reduce noise.
+  - Always report sampler statistics after each sample.
 
 ## v1.1.15